Split SOSE tasks/subtasks into separate files
A few small changes to image names, etc. have been included here.
